refactor: confirmation of pament
