+++ src/widgets/profileSettings/components/accountManagement/AccountManagement.tsx
@@ -4,17 +4,13 @@ import s from './AccountManagement.module.scss'
 import { SubscriptionSelector } from './SubscriptionSelector'
 import { CurrentSubscription } from './currentSubscription'
 import { useSearchParams } from 'next/navigation'
-import ConfirmationModal from './ConfirmationModal/ConfirmationModal'
 
 export const AccountManagement = () => {
   const [accountType, setAccountType] = useState('personal')
   const [autoRenew, setAutoRenew] = useState(true)
   const [subscriptionID, setSubscriptionID] = useState('')
-  const [confirmationModal, setConfirmationModal] = useState(false)
   const [showSubscription, setShowSubscription] = useState(false)
   const searchParams = useSearchParams()
-
-  const handleCloseConfimModal = () => setConfirmationModal(false)
 
   useEffect(() => {
     const success = searchParams.get('success')
@@ -23,7 +19,6 @@ export const AccountManagement = () => {
     if (id) setSubscriptionID(id)
 
     if (success === '1') {
-      setConfirmationModal(true)
       setShowSubscription(true)
     } else {
       setShowSubscription(false)
@@ -68,9 +63,6 @@ export const AccountManagement = () => {
           />
         </div>
         {accountType === 'business' && <SubscriptionSelector />}
-        {confirmationModal && (
-          <ConfirmationModal open={confirmationModal} onOpenChange={handleCloseConfimModal} />
-        )}
       </div>
     </section>
   )
